Added light refactoring to script.js, but more importantly moved different types of files into their own respective directories.
