+++ FoodAccessMap/css/style.css
@@ -1,0 +1,228 @@
+body,
+html {
+  height: 100%;
+  font-family: Arial;
+}
+
+#map {
+  width: 100%;
+  height: 100%;
+  z-index: 100;
+}
+#mapSearchContainer {
+  position: fixed;
+  top: 20px;
+  right: 40px;
+  height: 30px;
+  width: 180px;
+  z-index: 110;
+  font-size: 10pt;
+  color: #5d5d5d;
+  border: solid 1px #bbb;
+  background-color: #f8f8f8;
+}
+
+#searchInput {
+  width: 180px;
+}
+
+.geocoder-control-suggestion {
+  list-style-type: none;
+}
+.geocoder-control-suggestion:hover {
+  background-color: #DDD;
+}
+.geocoder-control-selected {
+  background-color: #AAA;
+}
+
+#results {
+  clear: left;
+}
+
+.entryDiv {
+  background-color: #88e1eb;
+  border: solid 1px #bbb;
+  border-radius: 25px;
+  padding-left: 10px;
+  padding-right: 10px;
+  padding-bottom: 5px;
+  padding-top: 5px;
+}
+
+.entryDiv > h2 {
+  font-size: 16px;
+  margin-bottom: 5px;
+}
+.entryDiv > p {
+  margin: 2px;
+}
+
+.accepts_icon {
+  border-radius: 5px;
+  font-weight: bold;
+  padding: 3px;
+}
+
+.snap {
+  background-color: darkgreen;
+  color: white;
+}
+
+.wic {
+  background-color: #ba55d3;
+}
+
+.foodbucks {
+  background-color: #32cd32;
+}
+
+.freshproduce {
+  background-color: orangered;
+}
+
+.fmnp {
+  background-color: darkred;
+  color: white;
+}
+
+input[type='text'] {
+  float: left;
+  padding: 6px;
+  border: solid 1px #bbb;
+  margin-top: 8px;
+  margin-right: 16px;
+  margin-bottom: 8px;
+  font-size: 17px;
+}
+
+.pointer {
+  position: absolute;
+  top: 86px;
+  left: 40vw;
+  z-index: 99999;
+}
+
+.descriptor-control {
+  padding: 5px 10px;
+  background-color: #f8f8f8;
+  color: #333;
+  font-size: 11px;
+  line-height: 18px;
+  border-radius: 5px;
+}
+
+.descriptorTitle {
+  font-size: 18px;
+  font-weight: bold;
+  color: #aa3333;
+}
+
+.descriptorBody {
+  font-size: 12px;
+}
+
+.descriptor-control:empty {
+  display: none;
+}
+
+.filter-control {
+  padding: 5px 10px;
+  background-color: #f8f8f8;
+  color: #333;
+  font-size: 11px;
+  line-height: 18px;
+  border-radius: 5px;
+}
+
+.filterTitle {
+  font-size: 18px;
+  font-weight: bold;
+  color: #aa3333;
+}
+
+.filterBody {
+  font-size: 12px;
+}
+
+.sourceOrg {
+  font-size: 10px;
+  color: #999;
+}
+
+.featureName {
+  font-size: 16px;
+  font-weight: bold;
+  color: #a83;
+}
+
+.descriptionTitle {
+  width: auto;
+  border-top: 1px solid #999;
+}
+.locationDescription {
+  margin-top: 5px;
+}
+.locationAddress {
+  margin-top: 5px;
+  color: #999;
+}
+
+.legend {
+  padding: 6px 8px;
+  font: 14px Arial, Helvetica, sans-serif;
+  background: white;
+  background: rgba(255, 255, 255, 0.8);
+  /*box-shadow: 0 0 15px rgba(0, 0, 0, 0.2);*/
+  /*border-radius: 5px;*/
+  line-height: 24px;
+  color: #555;
+}
+.legend h4 {
+  text-align: center;
+  font-size: 16px;
+  margin: 2px 12px 8px;
+  color: #777;
+}
+
+.legend span {
+  position: relative;
+  bottom: 3px;
+}
+
+.legend i {
+  width: 18px;
+  height: 18px;
+  float: left;
+  margin: 0 8px 0 0;
+  opacity: 0.7;
+}
+
+.legend i.icon {
+  background-size: 18px;
+  background-color: rgba(255, 255, 255, 1);
+}
+
+.togglepad {
+  margin-top: 10px;
+}
+/* @media screen and (orientation: landscape) {
+  #rotate-device-warning {
+    display: none;
+  }
+}
+
+@media screen and (orientation: portrait) {
+  #rotate-device-warning {
+    display: block;
+    width: 100%;
+    background-color: lightyellow;
+    padding: 10px;
+    border: 2px solid red;
+  }
+}
+@media screen and (orientation: portrait) and (min-device-width: 320px) {
+  .sidebar {
+    display: none;
+  }
+} */
