+++ FoodAccessMap/index.html
@@ -26,9 +26,9 @@
     <script src="https://unpkg.com/esri-leaflet@2.3.3/dist/esri-leaflet-debug.js"></script>
 
     <script src="https://unpkg.com/esri-leaflet-geocoder@2.3.2/dist/esri-leaflet-geocoder-debug.js"></script>
-    <link rel="stylesheet" href="./style.css">
-    <link rel="stylesheet" href="./leaflet-sidebar.css" />
-    <link rel="stylesheet" href="./introjs.css">
+    <link rel="stylesheet" href="css/style.css">
+    <link rel="stylesheet" href="js/leaflet-sidebar.css" />
+    <link rel="stylesheet" href="css/introjs.css">
 
 </head>
 
@@ -263,10 +263,10 @@
     <script src="https://gitcdn.github.io/bootstrap-toggle/2.2.2/js/bootstrap-toggle.min.js"></script>
     <script src='https://cdn.jsdelivr.net/npm/papaparse@5.3.0/papaparse.min.js'></script>
     <script src='https://unpkg.com/leaflet-knn@0.1.0/leaflet-knn.js'></script>
-    <script src="./leaflet-sidebar.js"></script>
-    <script src="./bootstrap-geocoder.js"></script>
-    <script src="./intro.js"></script>
-    <script src="./script.js"></script>
+    <script src="js/leaflet-sidebar.js"></script>
+    <script src="js/bootstrap-geocoder.js"></script>
+    <script src="js/intro.js"></script>
+    <script src="js/script.js"></script>
 
 </body>
